Perform resampling on float values
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/branches/libsidplayfp-experimental@2742 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/builders/residfp-builder/residfp/resample/TwoPassSincResampler.h
@@ -61,12 +61,12 @@ public:
         return new TwoPassSincResampler(clockFrequency, samplingFrequency, highestAccurateFrequency, intermediateFrequency);
     }
 
-    bool input(int sample) override
+    bool input(float sample) override
     {
         return s1->input(sample) && s2->input(s1->output());
     }
 
-    int output() const override
+    float output() const override
     {
         return s2->output();
     }
